pyproject.toml
[project]
name = "anth2oai"
version = "0.1.0"
description = "Anthropic to OpenAI API Wrapper - Use Claude models with OpenAI-compatible interface"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
keywords = ["anthropic", "openai", "claude", "api", "wrapper", "llm", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.54.0",
    "loguru>=0.7.0",
    "typing-extensions>=4.0.0",
    "fastapi[standard]>=0.115.11",
    "uvicorn>=0.34.0",
    "gunicorn>=23.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "uuid>=1.30",
    "tortoise-orm>=0.21.0",
    "aiosqlite>=0.19.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/anth2oai"
Repository = "https://github.com/yourusername/anth2oai"
Documentation = "https://github.com/yourusername/anth2oai#readme"
Issues = "https://github.com/yourusername/anth2oai/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["anth2oai"]

[tool.hatch.build.targets.sdist]
include = [
    "/anth2oai",
    "/README.md",
    "/LICENSE",
]
